requirements.txt
fastapi==0.103.2
google-cloud-aiplatform==1.34.0
grpcio-status==1.59.0
gunicorn==21.2.0
langchain==0.0.310
pip==22.3.1
pipdeptree==2.13.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
squareup==31.0.0.20230925
uvicorn==0.23.2
wheel==0.38.4
